Enum amethyst_animation::SpriteRenderPrimitive

pub enum SpriteRenderPrimitive {
    SpriteSheet(u64),
    SpriteIndex(usize),
}

Sampler primitive for SpriteRender animations Note that sprites can only ever be animated with Step, or a panic will occur.

Variants
SpriteSheet(u64)SpriteIndex(usize)