Function amethyst_animation::get_animation_set

pub fn get_animation_set<'a, I, T>(
    controls: &'a mut WriteStorage<'_, AnimationControlSet<I, T>>, 
    entity: Entity
) -> Option<&'a mut AnimationControlSet<I, T>> where
    I: Send + Sync + 'static,
    T: AnimationSampling, 

Get the animation set for an entity. If none exists, one will be added. If entity is invalid, (eg. removed before) None will be returned.

Type parameters:

I: identifier type for running animations, only one animation can be run at the same time with the same id

T: the component type that the animation applies to