Function amethyst_animation::pause_animation

pub fn pause_animation(
    controls: &mut WriteStorage<AnimationControl>, 
    animation: &Handle<Animation>, 
    entity: Entity
)

Pause the running animation on the given entity.

Parameters:

controls: animation control storage in the world.

animation: handle to the animation to run

entity: entity the animation is running on. Must either have an AnimationHierarchy that matches the Animation, or only refer to a single node, else the animation will not be run.