Struct amethyst_animation::AnimationBundle

pub struct AnimationBundle<'a> { /* fields omitted */ }

Bundle for a complete animation setup including sampler interpolation and animation control.

This will also add SamplingBundle, because it is a dependency of this bundle.

Will add AnimationControlSystem with name animation_control_system.

Methods

impl<'a> AnimationBundle<'a>

pub fn new() -> Self

Create a new animation bundle

pub fn with_dep(self, dep: &'a [&'a str]) -> Self

Set dependencies for the AnimationControlSystem