Struct amethyst_animation::AnimationControlSystem

pub struct AnimationControlSystem<I, T> where
    I: Eq + Hash,  { /* fields omitted */ }

System for setting up animations, should run before SamplerInterpolationSystem.

Will process all active AnimationControl + AnimationHierarchy, and do processing of the animations they describe. If an animation only targets a single node/entity, there is no need for AnimationHierarchy.

Type parameters:

I: identifier type for running animations, only one animation can be run at the same time with the same id

T: the component type that the animation should be applied to

Methods

impl<I, T> AnimationControlSystem<I, T> where
I: Eq + Hash,

pub fn new() -> Self