Struct amethyst_animation::AnimationSet

pub struct AnimationSet<I, T> where
    I: Eq + Hash,
    T: AnimationSampling,  {
    pub animations: FnvHashMap<I, Handle<Animation<T>>>,
}

Attaches to an entity that have animations, with links to all animations that can be run on the entity. Is not used directly by the animation systems, provided for convenience.

Type parameters:

T: the component type that the animation should be applied to

Fields
animations: FnvHashMap<I, Handle<Animation<T>>>
Methods

impl<I, T> AnimationSet<I, T> where
I: Eq + Hash,
T: AnimationSampling,

pub fn new() -> Self

Create

pub fn insert(&mut self, id: I, handle: Handle<Animation<T>>)

Insert an animation in the set

pub fn get(&self, id: &I) -> Option<&Handle<Animation<T>>>

Retrieve an animation handle from the set