Trait amethyst_animation::Interpolate

pub trait Interpolate<T> {
    fn interpolate(
        &self, 
        input: f32, 
        inputs: &[f32], 
        outputs: &[T], 
        normalize: bool
    ) -> T;
}

Interpolate over data set of the given type.

Required Methods

fn interpolate(
&self,
input: f32,
inputs: &[f32],
outputs: &[T],
normalize: bool
) -> T

Interpolation function, f(input) -> T

Parameters

input: the input value to the function

inputs: list of discrete input values for each keyframe

outputs: list of output values to interpolate between, note that this data set is tied to the interpolation function, and there is no guarantee or requirement that it is the same size as the inputs.

normalize: if true, normalize the interpolated value before returning it