Crate amethyst_animation

Structs

Animation
Defines a single animation. Defines relationships between the node index in AnimationHierarchy and a Sampler handle. If the animation only targets a single node index, AnimationHierarchy is not required.

AnimationBundle
Bundle for a complete animation setup including sampler interpolation and animation control.

AnimationControl
Attaches to an entity, to control what animations are currently active

AnimationControlSystem
System for setting up animations, should run before SamplerInterpolationSystem.

AnimationHierarchy
Defines the hierarchy of nodes that a single animation can control. Attach to the root entity that an animation can be defined for. Only required for animations which targets more than a single node.

AnimationSet
Attaches to an entity that have animations, with links to all animations that can be run on the entity. Is not used directly by the animation systems, provided for convenience.

Sampler
Sampler defines a single animation for a single attribute of the LocalTransform of the entity it is attached to.

SamplerControl
Run the sampler on the attached entity

SamplerControlSet
Sampler control set, containing optional samplers for each of the possible channels.

SamplerInterpolationSystem
System for interpolating active samplers.

SamplingBundle
Bundle for only the sampler interpolation.

Enums

AnimationCommand
Animation command

AnimationOutput
The actual animation data for a single attribute

ControlState
State of animation

EndControl
Control handling of animation/sampler end

InterpolationType
Supported interpolation functions

RestState
The rest state for a single attribute

Traits

Interpolate
Interpolate over data set of the given type.

InterpolationFunction
Trait used if an outside user wants to supply their own interpolation function

Functions

pause_animation
Pause the running animation on the given entity.

play_animation
Play a given animation on the given entity.

toggle_animation
Toggle the state between paused and running for the given animation on the given entity.

Type Definitions

AnimationProcessor
Asset storage processor for Animation

SamplerProcessor
Asset storage processor for Sampler