Struct amethyst_animation::SamplerControlSet

pub struct SamplerControlSet<T> where
    T: AnimationSampling,  {
    pub samplers: Vec<SamplerControl<T>>,
}

Sampler control set, containing a set of sampler controllers for a single component.

Have support for multiple samplers per channel, will do linear blending between all active samplers. The target component specifies if it can be blended, if it can't, the last added sampler wins.

Type parameters:

T: the component type that the sampling should be applied to

Fields
samplers: Vec<SamplerControl<T>>
Methods

impl<T> SamplerControlSet<T> where
T: AnimationSampling,

pub fn add_control(&mut self, control: SamplerControl<T>)

Set channel control

pub fn clear(&mut self, control_id: u64)

Clear sampler controls for the given animation

pub fn is_empty(&self) -> bool

Check if set is empty

pub fn abort(&mut self, control_id: u64)

Abort control set

pub fn pause(&mut self, control_id: u64)

Pause control set

pub fn unpause(&mut self, control_id: u64)

Unpause control set

pub fn set_rate_multiplier(&mut self, control_id: u64, rate_multiplier: f32) where
T: AnimationSampling,

Update rate multiplier

pub fn set_input(&mut self, control_id: u64, input: f32) where
T: AnimationSampling,

Forcibly set the input value (point of interpolation)

pub fn check_termination(&self, control_id: u64) -> bool

Check if a control set can be terminated

pub fn step(
&mut self,
control_id: u64,
samplers: &AssetStorage<Sampler<T::Primitive>>,
direction: &StepDirection
)

Step animation

pub fn set_blend_weight(
&mut self,
control_id: u64,
channel: &T::Channel,
blend_weight: f32
)

Set blend weight for a sampler

pub fn get_running_duration(&self, control_id: u64) -> Option<f32>

Get the max running duration of the control set